# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reflex delivery coordination API for Kenyan retailers
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/activity:
    get:
      operationId: listActivity
      tags: [dashboard]
      summary: List recent delivery activity
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Recent status activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
  /v1/sync:
    post:
      operationId: syncOfflineEvents
      tags: [riders]
      summary: Sync offline status events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncInput"
      responses:
        "200":
          description: Sync results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
components:
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: number
        minimum: 1
        maximum: 100
        default: 50
...
  schemas:
...
    DeliveryStatus:
      type: string
      enum: [pending, assigned, picked_up, delivered, cancelled]
    DeliveryRequest:
      type: object
      required: [id, reference, customerName, customerPhone, address, itemDescription, status, version, createdAt, updatedAt]
      properties:
        id: { type: string }
        reference: { type: string }
        customerName: { type: string }
        customerPhone: { type: string }
        address: { type: string }
        itemDescription: { type: string }
        status: { $ref: "#/components/schemas/DeliveryStatus" }
        riderId: { type: ["string", "null"] }
        riderName: { type: ["string", "null"] }
        version: { type: number }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        proof: { $ref: "#/components/schemas/ProofOfDelivery" }
...
    ProofOfDelivery:
      type: object
      required: [recipientName, verifiedAt]
      properties:
        recipientName: { type: string }
        signatureData: { type: ["string", "null"] }
        photoUrl: { type: ["string", "null"] }
        verifiedAt: { type: string, format: date-time }
...
    ActivityItem:
      type: object
      required: [id, deliveryId, reference, status, actorName, createdAt]
      properties:
        id: { type: string }
        deliveryId: { type: string }
        reference: { type: string }
        status: { $ref: "#/components/schemas/DeliveryStatus" }
        actorName: { type: string }
        createdAt: { type: string, format: date-time }
        note: { type: ["string", "null"] }
    SyncInput:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/OfflineEvent"
    OfflineEvent:
      type: object
      required: [deliveryId, status, version, clientEventId]
      properties:
        deliveryId: { type: string }
        status: { $ref: "#/components/schemas/DeliveryStatus" }
        version: { type: number }
        clientEventId: { type: string }
    SyncResult:
      type: object
      required: [accepted, duplicate, rejected, processedEventIds]
      properties:
        accepted: { type: number }
        duplicate: { type: number }
        rejected: { type: number }
        processedEventIds:
          type: array
          items: { type: string }
        deliveries:
          type: array
          items: { $ref: "#/components/schemas/DeliveryRequest" }
...